requirements.txt
aiogram==2.19
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pytz==2021.3
typing_extensions==4.1.1
yarl==1.7.2

loguru~=0.6.0
pydantic~=1.9.0
python-dotenv~=0.19.2